fix some problems

+++ app/thesportsdb_parser/utils.py
@@ -17,12 +17,13 @@ async def save_img_to_folder(sv_path, cur_img, title):
         tmp_path = folder + title.replace(" ", "_").replace("/", "_").replace("'", "_").lower() + ext
         new_path = tmp_path.replace(ext, ".webp")
         if not os.path.exists(new_path):
-            async with aiohttp.ClientSession() as session:
-                async with session.get(cur_img) as res:
-                    if res.status == 200:
-                        f = await aiofiles.open(tmp_path, mode='wb')
-                        await f.write(await res.read())
-                        await f.close()
+            if not os.path.exists(tmp_path):
+                async with aiohttp.ClientSession() as session:
+                    async with session.get(cur_img) as res:
+                        if res.status == 200:
+                            f = await aiofiles.open(tmp_path, mode='wb')
+                            await f.write(await res.read())
+                            await f.close()
 
                         cmd = ("cwebp -q " + str(100) + " " + tmp_path + " -o " + tmp_path[:tmp_path.index(".") - 4] + ".webp")
